xhash_get

+++ projects/helpers.c/src/hash/types.h
@@ -1,6 +1,6 @@
 #pragma once
 
-typedef size_t (hash_fn_t)(const void *key);
+typedef uint64_t (hash_fn_t)(const void *key);
 
 typedef struct hash_t hash_t;
 typedef struct hash_entry_t hash_entry_t;
